YgoSet.
This code snippet introduces the YgoSet class, which is used to represent a Yu-Gi-Oh! card set. The YgoSet class contains properties such as setName, setCode, numOfCards, and tcgDate, representing the name, code, number of cards, and release date of the card set, respectively. This code snippet also includes annotations for Hive database integration using the HiveType and HiveField decorators.